// resources/js/types/navigation.ts
import type { InertiaLinkProps } from '@inertiajs/react';
import type { LucideIcon } from 'lucide-react';

export type BreadcrumbItem = {
    title: string;
    href: NonNullable<InertiaLinkProps['href']>;
};

export type NavItem = {
    title: string;
    href: NonNullable<InertiaLinkProps['href']>;
    icon?: LucideIcon | null;
    isActive?: boolean;
};

export type MenuNavigationItem = {
    id: number;
    title: string;
    href: string | null;
    icon: string | null;
    children: MenuNavigationItem[];
};